# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Swarm API",
    "version": "1.76.0",
    "description": "Multi-agent orchestration API for Claude Code, Codex, and Gemini CLI. Enables task distribution, agent communication, and service discovery.\n\nMCP tools are documented separately in [MCP.md](./MCP.md)."
  },
  "servers": [
    {
      "url": "http://localhost:3013",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key via Authorization: Bearer <API_KEY>"
      },
      "agentId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-ID",
        "description": "Agent UUID for agent-scoped operations"
      }
    },
...
  },
  "paths": {
...
    "/api/config/resolved": {
      "get": {
        "summary": "Get resolved config (merged global + agent + repo scopes)",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "agentId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "repoId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "includeSecrets",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved config entries"
          }
        }
      }
    },
    "/api/config/env-presence": {
      "get": {
        "summary": "Check which of the given env var keys are currently set in process.env (presence only, no values)",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "keys",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Map of key -> boolean (true iff set in process.env)"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/config/reload": {
      "post": {
        "summary": "Reload global swarm_config into process.env (override=true) and re-init integrations (Slack, GitHub, Linear, AgentMail)",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reload result"
          },
          "500": {
            "description": "Reload failed"
          }
        }
      }
    },
    "/api/config/{id}": {
      "get": {
        "summary": "Get a single config entry by ID",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "includeSecrets",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Config entry"
          },
          "404": {
            "description": "Config not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a config entry by ID (including legacy reserved rows for cleanup)",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Config deleted"
          },
          "404": {
            "description": "Config not found"
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "summary": "List config entries with optional filters",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "scope",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "scopeId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "includeSecrets",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of config entries"
          }
        }
      },
      "put": {
        "summary": "Create or update a config entry (reserved env-only keys are rejected)",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scope": {
                    "type": "string",
                    "enum": [
                      "global",
                      "agent",
                      "repo"
                    ]
                  },
                  "scopeId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "key": {
                    "type": "string",
                    "minLength": 1
                  },
                  "value": {},
                  "isSecret": {
                    "type": "boolean"
                  },
                  "envPath": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "description": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "scope",
                  "key"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Config entry upserted"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
...
}